Facial Emotion Recognition using CNN
Facial expression refers to the movement of facial muscles that carry emotions expressed by a
person. It provides information about the mental status of that person. Emotion is a mental
condition that the person goes through. It is what he feels inside as the reaction of the events that
take place around him. Information about a person’s mental condition lies in his facial condition
in many cases. Fluctuation in recognition rate among the classes is one of the issues for most of
the research as they have lower recognition rate to detect emotions like disgust and fear.